hydra/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
httpx
websockets
tenacity
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
pytest
pytest-asyncio
playwright
aiosqlite
typer
aioredis
babel
